requirements-file.txt
PyQt5==5.15.6
rich==10.12.0
azure-cognitiveservices-speech==1.19.0
openai==0.27.0
elevenlabs==0.2.24
pygame==2.1.2
soundfile==0.10.3.post1
mutagen==1.45.1
